test: Recoil 샘플코드 MainPage > dummyPage로 이동
