+++ src/pages/index.ts
@@ -1,3 +1,4 @@
 export { default as MainPage } from "./MainPage";
 export { default as MyPage } from "./MyPage";
 export { default as ReviewPage } from "./ReviewPage";
+export { default as DummyPage } from "./DummyPage";
